alzheimer-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0